Redo directory structure, update project update copywrite years
