+++ Package.swift
@@ -3,7 +3,7 @@
 
 //MIT License
 //
-//Copyright (c) 2019 Ryan Baumbach <github@ryan.codes>
+//Copyright (c) 2019-2022 Ryan Baumbach <github@ryan.codes>
 //
 //Permission is hereby granted, free of charge, to any person obtaining a copy
 //of this software and associated documentation files (the "Software"), to deal
@@ -25,10 +25,28 @@
 
 import PackageDescription
 
-let package = Package(name: "Utensils",
-                      platforms: [.iOS(.v10)],
-                      products: [.library(name: "Utensils", targets: ["Utensils"])],
-                      dependencies: [.package(url: "https://github.com/rbaumbach/Capsule", from: "0.2.0")],
-                      targets: [.target(name: "Utensils",
-                                        dependencies: ["Capsule"])],
-                      swiftLanguageVersions: [.v5])
+let package = Package(
+    name: "Utensils",
+    platforms: [
+        .iOS(.v11)
+    ],
+    products: [
+        .library(
+            name: "Utensils",
+            targets: ["Utensils"]
+        )
+    ],
+    dependencies: [
+        .package(
+            url: "https://github.com/rbaumbach/Capsule",
+            from: "0.2.0"
+        )
+    ],
+    targets: [
+        .target(
+            name: "Utensils",
+            dependencies: ["Capsule"]
+        )
+    ],
+    swiftLanguageVersions: [.v5]
+)
