+++ Sources/Utensils/Models/Directory.swift
@@ -1,6 +1,6 @@
 //MIT License
 //
-//Copyright (c) 2020 Ryan Baumbach <github@ryan.codes>
+//Copyright (c) 2020-2022 Ryan Baumbach <github@ryan.codes>
 //
 //Permission is hereby granted, free of charge, to any person obtaining a copy
 //of this software and associated documentation files (the "Software"), to deal
